Test: Expand component and unit test coverage
- Add integration tests for HTTP client with additional methods and
  options
- Add edge-case tests for socket lifecycle and I/O operations
- Extend DNS parser tests for error paths and boundary conditions
- Add updater and network interface utility tests
- Fix narrowing conversion warnings in test code

+++ test/unit/util/random_test.cpp
@@ -63,7 +63,7 @@ TEST(RandomEngineTest, ReasonableDistribution) {
     auto& eng = Utils::Random::engine();
     std::set<std::uint32_t> seen;
     for (int i = 0; i < 1000; ++i) {
-        seen.insert(eng());
+        seen.insert(static_cast<std::uint32_t>(eng()));
     }
     // With 1000 draws from a 32-bit space, virtually all values are
     // unique, but we just check at least 990 to avoid flakiness.
